.travis.yml,CI/travis/after_deploy: trigger other builds after libiio
Let's start a "pipeline", where the libiio build triggers a
`libad9361-iio`, which will trigger other builds as well.

We're getting to a point where the lack of automation is starting to hinder
us a bit and changing things in `libiio` are not visible along the way,
until it gets a bit late.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>